Fix type mismatch errors in test_apetag.cpp.

+++ tests/test_apetag.cpp
@@ -98,10 +98,10 @@ public:
     CPPUNIT_ASSERT(unsuccessful.contains("A"));
     CPPUNIT_ASSERT(unsuccessful.contains("MP+"));
 
-    CPPUNIT_ASSERT_EQUAL((unsigned int)2, tag.itemListMap().size());
+    CPPUNIT_ASSERT_EQUAL((size_t)2, tag.itemListMap().size());
     tag.addValue("VALID KEY", "Test Value 1");
     tag.addValue("INVALID KEY \x7f", "Test Value 2");
-    CPPUNIT_ASSERT_EQUAL((unsigned int)3, tag.itemListMap().size());
+    CPPUNIT_ASSERT_EQUAL((size_t)3, tag.itemListMap().size());
   }
 
   void testTextBinary()
